backend/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.1
aiohttp-retry==2.8.3 ; python_version >= '3.7'
aiosignal==1.3.2 ; python_version >= '3.9'
amqp==5.3.1 ; python_version >= '3.6'
annotated-types==0.7.0 ; python_version >= '3.8'
asgiref==3.8.1
asttokens==3.0.0 ; python_version >= '3.8'
attrs==24.3.0 ; python_version >= '3.8'
authy==2.2.6
autobahn==24.4.2 ; python_version >= '3.9'
automat==24.8.1 ; python_version >= '3.8'
backoff==2.2.1
billiard==4.2.1 ; python_version >= '3.7'
blinker==1.9.0 ; python_version >= '3.9'
brotli==1.1.0
bson==0.5.10
cachecontrol==0.14.2 ; python_version >= '3.8'
cachetools==5.5.0 ; python_version >= '3.7'
celery[redis]==5.4.0
certifi==2024.12.14 ; python_version >= '3.6'
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
channels[daphne]==4.0.0
channels-redis==4.2.1
charset-normalizer==3.4.1 ; python_version >= '3.7'
click==8.1.8 ; python_version >= '3.7'
click-didyoumean==0.3.1 ; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0 ; python_version >= '3.6'
colorlog==6.9.0
constantly==23.10.4 ; python_version >= '3.8'
crispy-bootstrap4==2022.1
cryptography==44.0.0 ; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
daphne==4.1.2
decorator==5.1.1 ; python_version >= '3.11'
discord-webhook==1.3.1
dj-database-url==2.2.0
django==4.0.10
django-allauth==0.62.1
git+https://github.com/TheSpaghettiDetective/django-channels-presence.git@24c15f229372a16c7e8d6e160de89f4817d96c3f#egg=django-channels-presence
django-cors-headers==4.5.0
django-crispy-forms==2.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-hijack==3.4.5
django-ipware==7.0.1
django-jstemplate==2.0.1
django-oauth-toolkit==2.3.0
django-qr-code==3.1.1
django-redis==5.4.0
django-safedelete==1.4.0
django-settings-export==1.2.1
django-simple-history==3.6.0
django-slick-reporting==1.3.1
django-webpack-loader==3.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
executing==2.1.0 ; python_version >= '3.8'
firebase-admin==6.6.0
frozenlist==1.5.0 ; python_version >= '3.8'
google-api-core[grpc]==2.24.0 ; platform_python_implementation != 'PyPy'
google-api-python-client==2.159.0 ; python_version >= '3.7'
google-auth==2.37.0 ; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1 ; python_version >= '3.7'
google-cloud-firestore==2.20.0 ; platform_python_implementation != 'PyPy'
google-cloud-storage==2.19.0 ; python_version >= '3.7'
google-crc32c==1.6.0 ; python_version >= '3.9'
google-resumable-media==2.7.2 ; python_version >= '3.7'
googleapis-common-protos==1.66.0 ; python_version >= '3.7'
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h2==4.1.0
hpack==4.0.0 ; python_full_version >= '3.6.1'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
hyperframe==6.0.1 ; python_full_version >= '3.6.1'
hyperlink==21.0.0
idna==3.10 ; python_version >= '3.5'
incremental==24.7.2 ; python_version >= '3.8'
inotify-simple==1.3.5
ipdb==0.13.13
ipython==8.31.0 ; python_version >= '3.11'
jedi==0.19.2 ; python_version >= '3.6'
jsonfield==3.1.0
jwcrypto==1.5.6 ; python_version >= '3.8'
kombu==5.4.2 ; python_version >= '3.8'
matplotlib-inline==0.1.7 ; python_version >= '3.8'
msgpack==1.1.0 ; python_version >= '3.8'
multidict==6.1.0 ; python_version >= '3.8'
newrelic==10.4.0
nplusone==1.0.0
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2 ; python_version >= '3.6'
packaging==24.2
parso==0.8.4 ; python_version >= '3.6'
pexpect==4.9.0 ; sys_platform != 'win32' and sys_platform != 'emscripten'
phonenumbers==8.13.53
pillow==11.1.0
priority==1.3.0
prompt-toolkit==3.0.43
propcache==0.2.1 ; python_version >= '3.9'
proto-plus==1.25.0 ; python_version >= '3.7'
protobuf==5.29.3 ; python_version >= '3.8'
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pushbullet.py==0.12.0
pyasn1==0.6.1 ; python_version >= '3.8'
pyasn1-modules==0.4.1 ; python_version >= '3.8'
pycparser==2.22 ; python_version >= '3.8'
pydantic==2.8.0
pydantic-core==2.20.0
pygments==2.19.1 ; python_version >= '3.8'
pyjwt==2.10.1
pyopenssl==25.0.0
pyparsing==3.2.1 ; python_version >= '3.1'
pytelegrambotapi==4.26.0
python-dateutil==2.9.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-ipware==3.0.0 ; python_version >= '3.7'
python-magic==0.4.27 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pytz==2024.2
redis==5.2.1
requests==2.32.3 ; python_version >= '3.8'
requests-toolbelt==1.0.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
segno==1.6.1 ; python_version >= '3.5'
sentry-sdk==2.20.0
service-identity==24.2.0
setuptools==75.8.0 ; python_version >= '3.9'
simplejson==3.19.3 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.17.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.3 ; python_version >= '3.8'
stack-data==0.6.3
traitlets==5.14.3 ; python_version >= '3.8'
twilio==9.4.3
twisted[http2,tls]==24.11.0
txaio==23.1.1 ; python_version >= '3.7'
typing-extensions==4.11.0
tzdata==2024.2 ; python_version >= '2'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==2.3.0 ; python_version >= '3.9'
vine==5.1.0 ; python_version >= '3.6'
wcwidth==0.2.13
websocket-client==1.8.0 ; python_version >= '3.8'
whitenoise[brotli]==6.8.2
yarl==1.18.3 ; python_version >= '3.9'
zope.interface==7.2
openai===1.57.2
instructor===1.5.1
langfuse==2.57.6